% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{actionMessage}
\alias{actionMessage}
\title{EventLogger Action Message}
\arguments{
\item{msg}{The text to display and show}

\item{prefix}{Default '[+]', the text to show before the message}

\item{color}{Default 'red', the color of the displayed text}

\item{\dots}{Passed to \link{message}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The \link{log} table, invisibly
}
\description{
EventLogger object method to present and record an emphatic message
}
\details{
\preformatted{
## Method Usage:
myObject$actionMessage( help=TRUE )

myObject$actionMessage(msg="No message provided!", prefix='[+]', color="red")

}

This is just a wrapper for \link{message} where \code{prefix='[+]'}
and \code{color="red"}
}
\examples{

el <- EventLogger()
el$actionMessage("Outer airlock door not responding")

# Show the log, nicely formatted:
el
}
\seealso{
\link{log}, \link{message}
}
