% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{colMap}
\alias{colMap}
\title{EventLogger Color Map}
\value{
A list of lists
}
\description{
Internal EventLogger field to manage colorizing parameters
}
\details{
This is a structured list holding color names and
\link[crayon]{crayon} functions. It helps manage selection of
relevant colorizing functions from a set of prameters.

\preformatted{
## NORMALLY YOU WILL NOT WANT TO ACCESS THIS FIELD DIRECTLY
## ALTERING IT MAY RESULT IN CODE INSTABILITY
}
}
\seealso{
\link{colorMap}, \link{colNameToFunc}, \link{colorize}
}
