% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{colNameToFunc}
\alias{colNameToFunc}
\title{Color Name to Crayon Function}
\arguments{
\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A list of lists
}
\description{
EventLogger object method to generate a map of color names to
crayon functions
}
\details{
\preformatted{
## Method Usage:
myObject$colNameToFunc( help=TRUE )

myObject$colNameToFunc( )
}

This list generated by this method is simply a lookup list-of-lists
used to turn names ("cyan", "black") into \link[crayon]{crayon}
colorizing functions. It is utilized by \link{colorize} to convert
the user's color parameters into the appropriate colorizing
functions.

The function is run only when \link{colMap} has not already been
set up. Afterwards the resulting list is stored in \link{colMap}
for cached retrieval.
}
\seealso{
\link{colMap}, \link{colorize}
}
