% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{colorize}
\alias{colorize}
\title{Colorize}
\arguments{
\item{msg}{Default "". A character vector of text to colorize}

\item{color}{Default NULL. Optional text (eg "yellow")
corresponding to the foreground color of the text}

\item{bgcolor}{Default NULL. Optional text (eg "silver")
corresponding to the background color of the text}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A character vector with ANSI color codes injected by
    \link[crayon]{crayon}
}
\description{
EventLogger object method to colorize a string
}
\details{
\preformatted{
## Method Usage:
myObject$colorize( help=TRUE )

myObject$colorize( msg="", color=NULL, bgcolor=NULL )
}

Takes a character vector and applies foregrand and/or background
color to it.
}
\examples{

el <- EventLogger()
x <- el$colorize(c("This", "That"), "green")
x
message(paste(x, collapse=" ... and ... "))
}
\seealso{
\link[crayon]{crayon}, \link{colNameToFunc},
    \url{https://en.wikipedia.org/wiki/ANSI_escape_code#Colors}
}
