% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{colorMap}
\alias{colorMap}
\title{Map Color}
\arguments{
\item{color}{Required, a string describing a color, eg
"magenta". Can also be a function reference}

\item{bg}{Default FALSE. If TRUE, will pick the background color
corresponding to the name.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A function from the \link[crayon]{crayon}
}
\description{
EventLogger object method to turn a string into a crayon color function
}
\details{
\preformatted{
## Method Usage:
myObject$colorMap( help=TRUE )

myObject$colorMap( color, bg=FALSE )
}

Takes a color name as input, returns a \link[crayon]{crayon}
function. This is primarily a utility function.
}
\examples{

el <- EventLogger()
f <- el$colorMap("red")
base::message(f("Quick Brown Fox"))
f <- el$colorMap("cyan", TRUE)
base::message(f("Lazy Dog"))
}
\seealso{
\link{colorize}, \link[crayon]{crayon},
    \link{colNameToFunc}
}
