% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{dateMessage}
\alias{dateMessage}
\title{EventLogger Date Message}
\arguments{
\item{msg}{The text to display and show}

\item{\dots}{Passed to \link{message}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The \link{log} table, invisibly
}
\description{
EventLogger object method to present and record a message with a date
}
\details{
\preformatted{
## Method Usage:
myObject$dateMessage( help=TRUE )

myObject$dateMessage(msg="No message provided!", ...)

}

This is just a wrapper for \link{message} where \code{datestamp=TRUE}
}
\examples{

el <- EventLogger()
el$dateMessage("Something auspicious just happened")

# Show the log, nicely formatted:
el
}
\seealso{
\link{log}, \link{message}
}
