% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{debugMessage}
\alias{debugMessage}
\title{EventLogger Debug Message}
\arguments{
\item{msg}{The text to display and show}

\item{prefix}{Default '[DEBUG]', the text to show before the message}

\item{color}{Default 'white', the color of the displayed text}

\item{bgcolor}{Default 'blue', the background color of the
displayed text}

\item{\dots}{Passed to \link{message}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The \link{log} table, invisibly
}
\description{
EventLogger object method to present and record a message for debugging
}
\details{
\preformatted{
## Method Usage:
myObject$debugMessage( help=TRUE )

myObject$debugMessage(msg="No message provided!", prefix='[DEBUG]',
                      color="white", bgcolor="blue")

}

This is just a wrapper for \link{message} where
\code{prefix='[DEBUG]'}, \code{color="white"} and
\code{bgcolor="blue"}
}
\examples{

el <- EventLogger()
for (cv in 1:9) {
   el$debugMessage(c("Chevron", cv, "locked ..."))
}

# Show the log, nicely formatted:
el
}
\seealso{
\link{log}, \link{message}
}
