% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{err}
\alias{err}
\title{EventLogger Error Message}
\arguments{
\item{msg}{The text to display and show}

\item{prefix}{Default '[ERROR]', the text to show before the message}

\item{color}{Default 'red', the color of the displayed text}

\item{bgcolor}{Default 'yellow', the background color of the}

\item{\dots}{Passed to \link{message}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The \link{log} table, invisibly
}
\description{
EventLogger object method to present and record an error message
}
\details{
\preformatted{
## Method Usage:
myObject$err( help=TRUE )

myObject$err(msg="No message provided!", prefix='[ERROR]',
             color="red", bgcolor="yellow",...)

}

This is just a wrapper for \link{message} where
\code{prefix='[ERROR]'}, \code{color="red"} and
\code{bgcolor="yellow"}
}
\examples{

el <- EventLogger()
el$err("Weight is not defined")
# Show the log, nicely formatted:
el
}
\seealso{
\link{log}, \link{message}
}
