% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{EvLogObj}
\alias{EvLogObj}
\title{EventLogger Object}
\value{
An empty field if not set, or an EventLogger-compliant object
}
\description{
Internal EventLogger field pointing to another object that holds a
shared \link{log}
}
\details{
This package was designed to be inherited by other ReferenceClass
objects, and was also designed such that many objects could share a
common log. An object that inherits (contains) EventLogger can
choose to reference a different object with this field. If so,
operations that would normally alter the EventLogger fields will
instead act on the referenced object.

The other object is defined by the \code{log} parameter on creation.

\preformatted{
## NORMALLY YOU WILL NOT WANT TO ACCESS THIS FIELD DIRECTLY
## ALTERING IT MAY RESULT IN CODE INSTABILITY
}
}
