% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{fieldDescriptions}
\alias{fieldDescriptions}
\title{Field Descriptions}
\arguments{
\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A list of character strings
}
\description{
A list of brief descriptions for each object field
}
\details{
\preformatted{
## Method Usage:
myObject$fieldDescriptions( help=TRUE )

myObject$fieldDescriptions( )
}

This method returns a simple list of descriptive text for each
object field. It is designed to help the user understand the role
of each field.
}
\examples{

myEL <- EventLogger()
myEL$fieldDescriptions()

}
\seealso{
\link{help}
}
