% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{log}
\alias{log}
\title{EventLogger Log}
\value{
A data.table
}
\description{
Internal EventLogger field holding the data.table of events
}
\details{
A \link{data.table} (data.frame-compliant) table with two columns:
\code{$Date} (which is automatically populated at time of event)
and \code{$Message}

This field is intended to be accessed, but alterations (additions)
should be performed using the \link{message} functions
}
\examples{

el <- EventLogger( )
el$message("An entry")
el$message("A second entry")
el$log
}
\seealso{
\link{message}, \link{showLog}
}
