% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{logText}
\alias{logText}
\title{Log Text}
\arguments{
\item{widthDefault}{70% of the \code{option('width')}. This is the
length to be used when strwrap()ing the event text}

\item{relative}{- Default TRUE, will show the time elapsed between
events. If FALSE will show absolute time stamps}

\item{pad}{- Default 11, character padding used to align the time
column}

\item{n}{Default 0, if greater will only show that number of most
recent events}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\description{
EventLogger object method to generate pretty-formatted text of the log
}
\details{
\preformatted{
## Method Usage:
myObject$logText( help=TRUE )

myObject$logText( width=0.7 * getOption("width"),
                  relative=TRUE, pad=11, n=0 )
}

This method will parse the \link{log} data.table and generate a
human-friendly, colorized table of events. The left column will
report the time difference between log events, intending to help
gauge how long individual events are taking. The right column will
be event message text, strwrap()'ed to the user's \code{width} option.

This method will return the 'raw' character strings of the text. In
general you will likely wish to instead use \link{showLog} to have
the message shown in a way such that the ANSI color codes are
properly evaluated for display.
}
\examples{

el <- EventLogger()
el$message("A sample message")
el$logText()
}
\seealso{
\link{showLog}, \link{message}, \link{log}
}
