% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{showLog}
\alias{showLog}
\title{Show Log}
\arguments{
\item{...}{Will be passed to \link{logText}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\description{
EventLogger object method to pretty-print the event log
}
\details{
\preformatted{
## Method Usage:
myObject$showLog( help=TRUE )

myObject$showLog( ... )
}

A simple wrapper for logText(), the contents of which are displayed
in the shell using \code{cat}. This method is auto-invoked if an
EventLogger object is evaluated in the shell (and there are no
other ReferenceClass classes contained in the object that would
take precedence)
}
\examples{

el <- EventLogger()
el$message("A sample message")
el$showLog()
}
\seealso{
\link{logText}
}
