% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{tidyTime}
\alias{tidyTime}
\title{Tidy Time}
\arguments{
\item{x}{Default NULL. Expects a single numeric value in seconds}

\item{pad}{Default 0, a minimum width that the final string should
occupy}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A character string with ANSI color codes injected by
    \link[crayon]{crayon}
}
\description{
EventLogger object method to pretty-format a time difference
}
\details{
\preformatted{
## Method Usage:
myObject$tidyTime( help=TRUE )

myObject$tidyTime( (x=NULL, pad=0 )
}

Takes a character vector and applies foregrand and/or background
color to it. This method is used by \link{showLog} to highlight
shorter or longer time frames.
}
\examples{

el <- EventLogger()
base::message( el$tidyTime(0.00013) )
base::message( el$tidyTime(35232) )
}
\seealso{
\link{showLog}, \link{logText}
}
