% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{useCol}
\alias{useCol}
\title{EventLogger Use Color Flag}
\value{
A single logical value
}
\description{
Internal EventLogger field indicating if output should be colorized
}
\details{
A logical flag, if TRUE it indicates that messaeges should be
colorized with the \link[crayon]{crayon} package.

\preformatted{
## NORMALLY YOU WILL NOT WANT TO ACCESS THIS FIELD DIRECTLY
## Instead, use the \link{useColor} method to check and alter the value
}
}
\examples{

el <- EventLogger( )
el$err("Something is wrong!")
el$useColor(FALSE)
el$err("It' still wrong!")
}
\seealso{
\link{useColor}, \link{message}
}
