% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{vb}
\alias{vb}
\title{EventLogger Verbosity Flag}
\value{
A logical value
}
\description{
Internal EventLogger field holding the data.table of events
}
\details{
A logical flag setting if EventLogger is verbose or not.

\preformatted{
## NORMALLY YOU WILL NOT WANT TO ACCESS THIS FIELD DIRECTLY
## Instead, use the \link{verbose} method to check and alter the value
}
}
\examples{

el <- EventLogger( )
el$actionMessage("HELLO")
el$verbose( FALSE )
el$message("please no shouting")
el
}
\seealso{
\link{message}
}
