% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{verbose}
\alias{verbose}
\title{EventLogger Verbosity}
\arguments{
\item{newval}{Default NULL. If provided and can be made logical,
will set the flag. Inability to cast as logical will emit a
non-fatal error.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A single logical value, invisibly
}
\description{
EventLogger object method to get/set verbosity flag
}
\details{
\preformatted{
## Method Usage:
myObject$verbose( help=TRUE )

myObject$verbose( newval=NULL )
}

Gets or sets the flag determining if messages should be displayed
to the terminal, or just logged.
}
\examples{

el <- EventLogger()
el$err("Please be aware that something has gone wrong")
el$verbose(FALSE)
el$err("Another problem! But not on your screen. You'll need to check the log")
el
}
\seealso{
\link{message}, \link{colorize}, \link{vb}
}
